+++ packages/cli/src/codegen/schema.test.ts
@@ -1,3 +1,4 @@
+import assert from 'assert';
 import * as graphql from 'graphql/language';
 import prettier from 'prettier';
 import Schema from '../schema';
@@ -27,7 +28,6 @@ const testEntity = (generatedTypes: any[], expectedEntity: any) => {
     expectedMethod.static
       ? expect(method instanceof StaticMethod).toBe(true)
       : expect(method instanceof Method).toBe(true);
-
     expect(method.params).toStrictEqual(expectedMethod.params);
     expect(method.returnType).toStrictEqual(expectedMethod.returnType);
     expect(formatTS(method.body)).toBe(formatTS(expectedMethod.body));
@@ -535,5 +535,113 @@ describe('Schema code generator', () => {
         },
       ],
     });
+  });
+
+  test('get related method for WithBytes entity', () => {
+    const codegen = createSchemaCodeGen(`
+      type WithBytes @entity {
+        id: Bytes!
+        related: [RelatedBytes!]! @derivedFrom(field: "related")
+      }
+      
+      type RelatedBytes @entity {
+        id: ID!
+        related: WithBytes!
+      }
+    `);
+
+    const generatedTypes = codegen.generateTypes();
+
+    testEntity(generatedTypes, {
+      name: 'WithBytes',
+      members: [],
+      methods: [
+        {
+          name: 'constructor',
+          params: [new Param('id', new NamedType('Bytes'))],
+          returnType: undefined,
+          body: `
+          super()
+          this.set('id', Value.fromBytes(id));`,
+        },
+        {
+          name: 'save',
+          params: [],
+          returnType: new NamedType('void'),
+          body: `
+            let id = this.get('id');
+            assert(id != null, 'Cannot save WithBytes entity without an ID');
+            if (id) {
+              assert(id.kind == ValueKind.BYTES, \`Entities of type WithBytes must have an ID of type Bytes but the id '\${id.displayData()}' is of type \${id.displayKind()}\`);
+              store.set('WithBytes', id.toBytes().toHexString(), this);
+            }
+          `,
+        },
+
+        {
+          name: 'load',
+          static: true,
+          params: [new Param('id', new NamedType('Bytes'))],
+          returnType: new NullableType(new NamedType('WithBytes')),
+          body: `return changetype<WithBytes | null>(store.get('WithBytes', id.toHexString()));`,
+        },
+        {
+          name: 'loadInBlock',
+          static: true,
+          params: [new Param('id', new NamedType('Bytes'))],
+          returnType: new NullableType(new NamedType('WithBytes')),
+          body: `return changetype<WithBytes | null>(store.get_in_block('WithBytes', id.toHexString()));`,
+        },
+        {
+          name: 'get id',
+          params: [],
+          returnType: new NamedType('Bytes'),
+          body: `let value = this.get("id")
+          if (!value || value.kind == ValueKind.NULL) {
+            throw new Error("Cannot return null for a required field.")
+          } else {
+            return value.toBytes()
+          }
+          `,
+        },
+        {
+          name: 'set id',
+          params: [new Param('value', new NamedType('Bytes'))],
+          returnType: undefined,
+          body: `this.set('id', Value.fromBytes(value));`,
+        },
+        {
+          name: 'get related',
+          params: [],
+          returnType: new NamedType('RelatedBytesLoader'),
+          body: `return new RelatedBytesLoader('WithBytes', this.get('id')!.toBytes().toHexString(), 'related');`,
+        },
+        // Add any additional getters and setters for other fields if necessary
+      ],
+    });
+  });
+
+  test('no derived loader for interface', () => {
+    const codegen = createSchemaCodeGen(`
+    interface IExample {
+      id: ID! 
+      main: Main!
+      num: Int!
+    }
+    
+    type Example1 implements IExample @entity {
+      id: ID! 
+      main: Main!
+      num: Int!
+    }
+    
+    type Main @entity {
+      id: ID!
+      examples: [IExample!]! @derivedFrom(field: "main")
+    }
+`);
+
+    const generateDerivedLoaders = codegen.generateDerivedLoaders().filter(Boolean);
+    assert(generateDerivedLoaders.length === 0);
   });
 });