+++ packages/cli/src/command-helpers/data-sources.ts
@@ -1,5 +1,5 @@
-import yaml from 'js-yaml';
 import immutable from 'immutable';
+import yaml from 'js-yaml';
 import { loadManifest } from '../migrations/util/load-manifest';
 import Protocol from '../protocols';