+++ packages/cli/src/compiler/asc.ts
@@ -72,6 +72,6 @@ export const compile = ({ inputFile, global, baseDir, libs, outputFile }: Compil
 
   assemblyScriptCompiler(compilerArgs, compilerDefaults);
 
-  // only if compiler succeded, that is, when the line above doesn't throw
+  // only if compiler succeeded, that is, when the line above doesn't throw
   removeExitHandler(exitHandler);
 };