+++ packages/cli/src/debug.ts
@@ -5,10 +5,17 @@ debugFactory.formatters.L = immutableList => {
 };
 
 debugFactory.formatters.M = immutableMap => {
-  if (immutableMap.toMap != null) {
+  if (immutableMap?.toMap != null) {
     return JSON.stringify(immutableMap.toMap());
   }
 
+  if (typeof immutableMap?.toJS === 'function') {
+    return JSON.stringify(immutableMap.toJS());
+  }
+
+  if (typeof immutableMap === 'object') {
+    return JSON.stringify(immutableMap);
+  }
   return immutableMap;
 };