+++ packages/cli/tests/cli/validation.test.ts
@@ -2,6 +2,14 @@ import { cliTest } from './util';
 
 describe('Validation', () => {
   cliTest(
+    'Block handler filters',
+    ['codegen', '--skip-migrations'],
+    'validation/block-handler-filters',
+    {
+      exitCode: 0,
+    },
+  );
+  cliTest(
     'Invalid subgraph manifest',
     ['codegen', '--skip-migrations'],
     'validation/invalid-manifest',
@@ -52,6 +60,12 @@ describe('Validation', () => {
       exitCode: 0,
     },
   );
+  cliTest('2D arrays are valid', ['codegen', '--skip-migrations'], 'validation/2d-array-is-valid', {
+    exitCode: 0,
+  });
+  cliTest('3D arrays are valid', ['codegen', '--skip-migrations'], 'validation/3d-array-is-valid', {
+    exitCode: 0,
+  });
   cliTest(
     'Missing entity "id" field',
     ['codegen', '--skip-migrations'],